added dynamic sliders

+++ Rubricas_PCL/Asignatura/Estudiante/Reporte/Parcial.cs
@@ -4,10 +4,10 @@ namespace Rubricas_PCL
 {
 	public class Parcial
 	{
-		public float Nota { get; set; }
+		public double Nota { get; set; }
 		public string Name { get; set; }
 
-        public Parcial(string name, float nota)
+        public Parcial(string name, double nota)
         {
             this.Name = name;
             this.Nota = nota;
